🎨 Restructure folder structure  to original version, change paths in vite.config + package.json instead
